+++ ase/calculators/calculator.py
@@ -1,4 +1,3 @@
-from abc import ABC, abstractmethod
 import os
 import copy
 import subprocess
@@ -13,6 +12,7 @@ from ase.cell import Cell
 from ase.dft.kpoints import monkhorst_pack
 from ase.outputs import Properties, all_outputs
 from ase.utils import jsonable
+from ase.calculators.abc import GetPropertiesMixin
 
 
 class CalculatorError(RuntimeError):
@@ -441,44 +441,6 @@ class Parameters(dict):
         pathlib.Path(filename).write_text(self.tostring())
 
 
-class GetPropertiesMixin(ABC):
-    """Mixin class which provides get_forces(), get_stress() and so on.
-
-    Inheriting class must implement get_property()."""
-
-    @abstractmethod
-    def get_property(self, name, atoms=None, allow_calculation=True):
-        """Get the named property."""
-
-    def get_potential_energies(self, atoms=None):
-        return self.get_property('energies', atoms)
-
-    def get_forces(self, atoms=None):
-        return self.get_property('forces', atoms)
-
-    def get_stress(self, atoms=None):
-        return self.get_property('stress', atoms)
-
-    def get_stresses(self, atoms=None):
-        """the calculator should return intensive stresses, i.e., such that
-                stresses.sum(axis=0) == stress
-        """
-        return self.get_property('stresses', atoms)
-
-    def get_dipole_moment(self, atoms=None):
-        return self.get_property('dipole', atoms)
-
-    def get_charges(self, atoms=None):
-        return self.get_property('charges', atoms)
-
-    def get_magnetic_moment(self, atoms=None):
-        return self.get_property('magmom', atoms)
-
-    def get_magnetic_moments(self, atoms=None):
-        """Calculate magnetic moments projected onto atoms."""
-        return self.get_property('magmoms', atoms)
-
-
 class Calculator(GetPropertiesMixin):
     """Base-class for all ASE calculators.
 
@@ -594,6 +556,9 @@ class Calculator(GetPropertiesMixin):
 
         if not hasattr(self, 'name'):
             self.name = self.__class__.__name__.lower()
+
+        if not hasattr(self, 'get_spin_polarized'):
+            self.get_spin_polarized = self._deprecated_get_spin_polarized
 
     @property
     def directory(self) -> str:
@@ -876,7 +841,12 @@ class Calculator(GetPropertiesMixin):
         else:
             return stress
 
-    def get_spin_polarized(self):
+    def _deprecated_get_spin_polarized(self):
+        msg = ('This calculator does not implement get_spin_polarized().  '
+               'In the future, calc.get_spin_polarized() will work only on '
+               'calculator classes that explicitly implement this method or '
+               'inherit the method via specialized subclasses.')
+        warnings.warn(msg, FutureWarning)
         return False
 
     def band_structure(self):